adds only delete your own comment logic

+++ src/Comment/comment.controller.ts
@@ -2,12 +2,16 @@ import {
   Body,
   Controller,
   Delete,
+  HttpException,
+  HttpStatus,
   Param,
   Post,
   UseGuards,
 } from '@nestjs/common';
 import { Comment, Prisma } from '@prisma/client';
 import { JwtAuthGuard } from 'src/auth/jwt.guard';
+import { jwtUser } from 'src/Decorators/jwtUser.Decorator';
+import { ParamPipe } from 'src/Pipes/param.pipe';
 import { CommentService } from './comment.service';
 import { CreateCommentDto } from './CommentDto/createComment.dto';
 import { DeleteCommentDto } from './CommentDto/DeleteComment.dto';
@@ -27,10 +31,21 @@ export class CommentController {
 
   @Delete(':id')
   async deleteComment(
-    @Param('id') id: string,
+    @jwtUser() jwtUser,
+    @Param('id', new ParamPipe('commentId')) id: string,
     @Body()
     deleteCommentDto: DeleteCommentDto,
   ) {
+    const isCommentOwned = await this.commentService.isCommentOwnedByUser(
+      jwtUser.id,
+      id,
+    );
+    if (isCommentOwned === false) {
+      throw new HttpException(
+        'This is not your comment to delete!',
+        HttpStatus.BAD_REQUEST,
+      );
+    }
     return this.commentService.deleteComment(deleteCommentDto);
   }
 }
